Implement navigation between Home and Product Detail screens and add recommendations
- Set up `NavHost` in `MainActivity` and configure dynamic routing for `ProductDetail` using `productID` parameters.
- Update `HomeContract` and `HomeRoute` to handle product click events, passing the selected product ID to the navigator.
- Implement `ProductDetailViewModel` to retrieve product details using `SavedStateHandle` and populate the UI state.
- Add `AlsoLike` component and `AlsoLikeProduct` domain model to display recommended products based on brand and category.
- Enhance `HomeScreen` and `ProductDetailScreen` with `CircularProgressIndicator` to handle `UiState.Loading` states.
- Refine `AdderButton` UI with improved padding, clipping, and `RoundedCornerShape`.
- Update `Product` component to be clickable and display the product ID for debugging.
- Expand `ProductModel` mock data to include comprehensive care instructions, materials, and icons.

+++ app/src/main/java/com/example/stylehub/presentation/home/HomeContract.kt
@@ -30,5 +30,5 @@ data class HomeState(
  **/
 data class HomeActions(
     val onClick: () -> Unit = {},
-    val onProductClick:() -> Unit={}
+    val onProductClick: (Int) -> Unit = {}
 )
